@@ -1,22 +1,46 @@
 const express = require("express");
 const router = express.Router()
+const jwt = require("jsonwebtoken");
+const { JWT_SECRET } = process.env;
+const { getUserById } = require("../db/models/user")
 
 
-router.get('/', (req, res, next) => {
-  res.send({
-    message: 'API is under construction!',
-  });
+router.use(async (req, res, next) => {
+  const prefix = 'Bearer ';
+  const auth = req.header('Authorization');
+  if (!auth) { 
+    next();
+  } else if (auth.startsWith(prefix)) {
+    const token = auth.slice(prefix.length);
+    
+    try {
+      const { id } = jwt.verify(token, JWT_SECRET);
+
+      if (id) {
+        req.user = await getUserById(id);
+        next();
+      }
+    } catch ({ name, message }) {
+      next({ name, message });
+    }
+  } else {
+    next({
+      name: 'AuthorizationHeaderError',
+      message: `Authorization token must start with ${ prefix }`
+    });
+  }
 });
 
 // place your routers here
 
+
+
 const usersRouter = require("./users")
-usersRouter.use("/users", usersRouter)
+router.use("/users", usersRouter)
+
 
 const pastriesRouter = require("./pastries")
-pastriesRouter.use("/pastries", pastriesRouter)
-
-
+router.use("/pastries", pastriesRouter);
 
 router.get('/health', (req, res, next) => {
   res.send({
