@@ -4,7 +4,8 @@ module.exports = {
   // for example, create a users.js file for a User model
   // and User: require('./user') here
   ...require("./user"),
-  ...require("./pastries")
+  ...require("./pastries"),
+  ...require("./cart")
 };
 
 // then, in your API, you'll require the appropriate model
