@@ -1,32 +1,107 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from "react";
 // getAPIHealth is defined in our axios-services directory index.js
 // you can think of that directory as a collection of api adapters
 // where each adapter fetches specific info from our express server's /api route
-import { getAPIHealth } from '../axios-services';
-import '../style/App.css';
+import { getAPIHealth } from "../axios-services";
+import "../style/App.css";
+import { Route, Routes, Link } from "react-router-dom";
+import Register from "./Register";
+
+import Pastries from "./Pastries";
+import { callApi } from "../api/utils";
+import Login from "./Login";
+import CreateForm from "./CreateForm";
 
 const App = () => {
-  const [APIHealth, setAPIHealth] = useState('');
+  const [APIHealth, setAPIHealth] = useState("");
+  const [error, setError] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [token, setToken] = useState("");
+  const [pastries, setPastries] = useState([]);
+  
+  console.log('error :>> ', error);
 
+  const fetchPastries = async () => {
+    const data = await callApi({
+      path: "/pastries"
+    })
+    setPastries(data);
+  };
+  
   useEffect(() => {
-    // follow this pattern inside your useEffect calls:
-    // first, create an async function that will wrap your axios service adapter
-    // invoke the adapter, await the response, and set the data
-    const getAPIStatus = async () => {
-      const { healthy } = await getAPIHealth();
-      setAPIHealth(healthy ? 'api is up! :D' : 'api is down :/');
-    };
-
-    // second, after you've defined your getter above
-    // invoke it immediately after its declaration, inside the useEffect callback
-    getAPIStatus();
+    fetchPastries();
   }, []);
-
+  
   return (
-    <div className="app-container">
-      <h1>Hello, World!</h1>
-      <p>API Status: {APIHealth}</p>
-    </div>
+      <div className="app-container">
+        <nav>
+          <ul>
+            <li>
+              <Link to="/register">Register</Link>
+            </li>
+            <li>
+              <Link to="login">Login</Link>
+            </li>
+            <li>
+            <Link to="/pastries">Pastries</Link>
+          </li>
+          <li>
+            <Link to="/admin">Admin</Link>
+          </li>
+          </ul>
+        </nav>
+        <Routes>
+          <Route
+            path="/register"
+            element={
+              <Register
+                setError={setError}
+                error={error}
+                email={email}
+                setEmail={setEmail}
+                password={password}
+                setPassword={setPassword}
+                setToken={setToken}
+              />
+            }
+          ></Route>
+          <Route
+            path="/login"
+            element={
+              <Login
+                setError={setError}
+                error={error}
+                setEmail={setEmail}
+                email={email}
+                setPassword={setPassword}
+                password={password}
+                setToken={setToken}
+              />
+            }
+          ></Route>
+          <Route
+          path="/pastries"
+          element={
+            <Pastries 
+              pastries={pastries}
+              setPastries={setPastries}
+            />
+          }
+      ></Route>
+      <Route
+          path="/admin"
+          element={
+            <CreateForm
+              pastries={pastries}
+              setPastries={setPastries}
+              token={token}
+            />
+          }
+      ></Route>
+        </Routes>
+      </div>
+
   );
 };
 
